refactor retry approach, handle batches correctly
