+++ rust/otap-dataflow/Cargo.toml
@@ -112,40 +112,3 @@ inherits = "release"
 debug = true  # Or 2 for full debug info
 strip = "none" # Keep symbols and debug info
 panic = "unwind"
-
-# Main crate package
-[package]
-name = "otap-dataflow"
-description = "OTAP dataflow library with shared traits"
-version.workspace = true
-authors.workspace = true
-repository.workspace = true
-license.workspace = true
-publish.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# Internal crates - removing circular dependencies
-# otap-df-otlp = { path = "crates/otlp" }
-# otap-df-otap = { path = "crates/otap" }
-otap-df-traits = { path = "crates/traits" }
-otap-df-config = { path = "crates/config" }
-otap-df-engine = { path = "crates/engine" }
-
-# External dependencies for trait implementations
-prost = "0.13.5"
-thiserror.workspace = true
-async-trait.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-tokio.workspace = true
-log.workspace = true
-
-[dev-dependencies]
-# For realistic testing with real telemetry data types
-otap-df-otlp = { path = "crates/otlp" }
-otap-df-otap = { path = "crates/otap" }
