+++ rust/otap-dataflow/crates/engine/Cargo.toml
@@ -20,8 +20,10 @@ otap-df-channel = { path = "../channel" }
 otap-df-config = { path = "../config" }
 
 thiserror = { workspace = true }
+serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 async-trait = { workspace = true }
+log = { workspace = true }
 
 socket2 = { version = "0.6.0", features = ["all"] }
