+++ rust/otap-dataflow/crates/otlp/Cargo.toml
@@ -13,7 +13,6 @@ rust-version.workspace = true
 workspace = true
 
 [dependencies]
-otap-dataflow = { path = "../.." }
 otap-df-traits = { path = "../traits" }
 otap-df-engine = { path = "../engine", features = ["testing"] }
 otap-df-channel = { path = "../channel" }
