--- rust/otap-dataflow/src/retry_processor.rs
@@ -1,876 +1,0 @@
-use async_trait::async_trait;
-use otap_df_engine::error::Error;
-use otap_df_engine::local::processor::{EffectHandler, Processor};
-use otap_df_engine::message::{ControlMsg, Message};
-use otap_df_traits::Retryable;
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::time::{Duration, Instant};
-
-/// Defines the control messages used by the retry processor for handling acknowledgments
-/// and negative acknowledgments from exporters.
-#[derive(Debug, Clone)]
-pub enum ControlMessage {
-    /// Acknowledgment that a message was successfully processed
-    Ack {
-        /// Message IDs that were successfully processed
-        msg_ids: Vec<String>,
-    },
-    /// Negative acknowledgment indicating processing failure
-    Nack {
-        /// Message IDs that failed to process
-        msg_ids: Vec<String>,
-        /// Details about the error that occurred
-        error: ErrorDetail,
-        /// Optional retry policy to override default behavior
-        retry_policy: Option<RetryPolicy>,
-    },
-}
-
-/// Retry policy configuration for failed messages
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RetryPolicy {
-    /// Delay in milliseconds before retrying
-    pub retry_delay_ms: u64,
-}
-
-/// Details about an error that occurred during message processing
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ErrorDetail {
-    /// Whether the error is recoverable and should be retried
-    pub recoverable: bool,
-    /// Human-readable description of the error
-    pub reason: String,
-}
-
-/// Configuration for the retry processor
-///
-/// This configuration provides comprehensive control over retry behavior and memory safety.
-///
-/// # Memory Safety Features
-///
-/// - **`max_pending_messages`**: Prevents unbounded memory growth by limiting the number of
-///   messages kept in the retry queue. When this limit is reached, new messages are still
-///   forwarded but won't be tracked for retry.
-///
-/// - **`cleanup_interval_secs`**: Periodic cleanup removes expired messages from the retry
-///   queue based on their deadlines, preventing memory leaks from messages that will never
-///   be retried.
-///
-/// # Example Usage
-///
-/// ```rust
-/// use otap_dataflow::retry_processor::RetryConfig;
-///
-/// // Production configuration with memory safety
-/// let config = RetryConfig {
-///     max_retries: 5,
-///     initial_retry_delay_ms: 1000,
-///     max_retry_delay_ms: 60000,
-///     backoff_multiplier: 2.0,
-///     max_pending_messages: 50000,  // Limit memory usage
-///     cleanup_interval_secs: 300,   // Cleanup every 5 minutes
-/// };
-/// ```
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RetryConfig {
-    /// Maximum number of retry attempts
-    pub max_retries: usize,
-    /// Initial retry delay in milliseconds
-    pub initial_retry_delay_ms: u64,
-    /// Maximum retry delay in milliseconds
-    pub max_retry_delay_ms: u64,
-    /// Backoff multiplier for exponential backoff
-    pub backoff_multiplier: f64,
-    /// Maximum number of pending messages to keep in memory
-    pub max_pending_messages: usize,
-    /// Interval for cleaning up expired messages
-    pub cleanup_interval_secs: u64,
-}
-
-impl Default for RetryConfig {
-    fn default() -> Self {
-        Self {
-            max_retries: 3,
-            initial_retry_delay_ms: 1000,
-            max_retry_delay_ms: 30000,
-            backoff_multiplier: 2.0,
-            max_pending_messages: 10000, // Default to 10,000 pending messages
-            cleanup_interval_secs: 60,   // Default to 60-second cleanup interval
-        }
-    }
-}
-#[derive(Debug, Clone)]
-struct PendingMessage<T> {
-    data: T,
-    retry_count: usize,
-    next_retry_time: Instant,
-    last_error: String,
-}
-
-/// The retry processor that handles failed message delivery with configurable retry logic
-pub struct RetryProcessor<T: Retryable> {
-    config: RetryConfig,
-    pending_messages: HashMap<u64, PendingMessage<T>>,
-    next_message_id: u64,
-    last_cleanup_time: Instant,
-}
-
-impl<T: Retryable> RetryProcessor<T> {
-    /// Creates a new retry processor with default configuration
-    #[must_use]
-    pub fn new() -> Self {
-        Self::with_config(RetryConfig::default())
-    }
-
-    /// Creates a new retry processor with the specified configuration
-    #[must_use]
-    pub fn with_config(config: RetryConfig) -> Self {
-        Self {
-            config,
-            pending_messages: HashMap::new(),
-            next_message_id: 1,
-            last_cleanup_time: Instant::now(),
-        }
-    }
-
-    /// Acknowledges successful processing of a message
-    fn acknowledge(&mut self, id: u64) {
-        let _ = self.pending_messages.remove(&id);
-    }
-
-    /// Handles negative acknowledgment of a message
-    async fn handle_nack(
-        &mut self,
-        id: u64,
-        reason: String,
-        _effect_handler: &mut EffectHandler<T>,
-    ) -> Result<(), Error<T>> {
-        if let Some(mut pending) = self.pending_messages.remove(&id) {
-            pending.retry_count += 1;
-            pending.last_error = reason;
-
-            if pending.retry_count <= self.config.max_retries {
-                // Calculate next retry time with exponential backoff
-                let delay_ms = (self.config.initial_retry_delay_ms as f64
-                    * self
-                        .config
-                        .backoff_multiplier
-                        .powi(pending.retry_count as i32 - 1))
-                .min(self.config.max_retry_delay_ms as f64) as u64;
-
-                pending.next_retry_time = Instant::now() + Duration::from_millis(delay_ms);
-                let _ = self.pending_messages.insert(id, pending);
-            } else {
-                // Max retries exceeded, log and drop the message
-                log::error!(
-                    "Message {} exceeded max retries ({}), dropping. Last error: {}",
-                    id,
-                    self.config.max_retries,
-                    pending.last_error
-                );
-            }
-        }
-        Ok(())
-    }
-
-    /// Processes pending retries that are ready to be sent
-    async fn process_pending_retries(
-        &mut self,
-        effect_handler: &mut EffectHandler<T>,
-    ) -> Result<(), Error<T>> {
-        let now = Instant::now();
-        let mut ready_messages = Vec::new();
-
-        // Find messages ready for retry
-        for (&id, pending) in &self.pending_messages {
-            if pending.next_retry_time <= now {
-                ready_messages.push((id, pending.data.clone()));
-            }
-        }
-
-        // Send ready messages
-        for (_id, data) in ready_messages {
-            effect_handler.send_message(data).await?;
-            // Keep the message in pending state - it will be removed on Ack or updated on Nack
-        }
-
-        Ok(())
-    }
-
-    /// Periodically cleans up expired messages from the pending queue
-    fn cleanup_expired_messages(&mut self) {
-        let now = Instant::now();
-        if now.duration_since(self.last_cleanup_time)
-            < Duration::from_secs(self.config.cleanup_interval_secs)
-        {
-            return; // Not time for cleanup yet
-        }
-
-        let expired_ids: Vec<u64> = self
-            .pending_messages
-            .iter()
-            .filter_map(|(&id, pending)| {
-                if let Some(deadline) = pending.data.deadline() {
-                    if now > deadline { Some(id) } else { None }
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        for id in expired_ids {
-            if self.pending_messages.remove(&id).is_some() {
-                log::warn!("Removed expired message with ID: {id}");
-            }
-        }
-
-        self.last_cleanup_time = now;
-    }
-
-    /// Adds a new message to the retry queue with an ID
-    fn add_message_for_retry(&mut self, data: T) -> Result<u64, String> {
-        if self.pending_messages.len() >= self.config.max_pending_messages {
-            return Err(format!(
-                "Retry queue is full (capacity: {}), cannot add message",
-                self.config.max_pending_messages
-            ));
-        }
-
-        let id = self.next_message_id;
-        self.next_message_id += 1;
-
-        let pending = PendingMessage {
-            data,
-            retry_count: 0,
-            next_retry_time: Instant::now(),
-            last_error: String::new(),
-        };
-
-        let _ = self.pending_messages.insert(id, pending);
-        Ok(id)
-    }
-}
-
-#[async_trait(?Send)]
-impl<T: Retryable> Processor<T> for RetryProcessor<T> {
-    async fn process(
-        &mut self,
-        msg: Message<T>,
-        effect_handler: &mut EffectHandler<T>,
-    ) -> Result<(), Error<T>> {
-        match msg {
-            Message::PData(data) => {
-                // Forward the data message and track it for potential retry
-                match self.add_message_for_retry(data.clone()) {
-                    Ok(_id) => {
-                        effect_handler.send_message(data).await?;
-                    }
-                    Err(error_msg) => {
-                        log::warn!("{error_msg}");
-                        // Still try to send the message even if we can't track it for retry
-                        effect_handler.send_message(data).await?;
-                    }
-                }
-                Ok(())
-            }
-            Message::Control(control_msg) => {
-                match control_msg {
-                    ControlMsg::Ack { id } => {
-                        self.acknowledge(id);
-                        Ok(())
-                    }
-                    ControlMsg::Nack { id, reason } => {
-                        self.handle_nack(id, reason, effect_handler).await
-                    }
-                    ControlMsg::TimerTick { .. } => {
-                        // Process pending retries on timer tick
-                        self.process_pending_retries(effect_handler).await?;
-                        // Also cleanup expired messages
-                        self.cleanup_expired_messages();
-                        Ok(())
-                    }
-                    ControlMsg::Config { config } => {
-                        // Update retry configuration if provided
-                        if let Ok(new_config) = serde_json::from_value::<RetryConfig>(config) {
-                            self.config = new_config;
-                        }
-                        Ok(())
-                    }
-                    ControlMsg::Shutdown { .. } => {
-                        // On shutdown, attempt to send all pending messages one final time
-                        let pending_ids: Vec<u64> = self.pending_messages.keys().cloned().collect();
-                        for id in pending_ids {
-                            if let Some(pending) = self.pending_messages.remove(&id) {
-                                let _ = effect_handler.send_message(pending.data).await;
-                            }
-                        }
-                        Ok(())
-                    }
-                }
-            }
-        }
-    }
-}
-
-impl<T: Retryable> Default for RetryProcessor<T> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use otap_df_engine::config::ProcessorConfig;
-    use otap_df_engine::message::{ControlMsg, Message};
-    use otap_df_engine::processor::ProcessorWrapper;
-    use otap_df_engine::testing::processor::TestRuntime;
-    use otap_df_otap::grpc::OTAPData;
-    use otap_df_otap::proto::opentelemetry::experimental::arrow::v1::BatchArrowRecords;
-    use otap_df_otlp::grpc::OTLPData;
-    use otap_df_otlp::proto::opentelemetry::collector::logs::v1::ExportLogsServiceRequest;
-    use otap_df_otlp::proto::opentelemetry::collector::metrics::v1::ExportMetricsServiceRequest;
-    use otap_df_otlp::proto::opentelemetry::collector::trace::v1::ExportTraceServiceRequest;
-    use std::time::Duration;
-
-    // Helper functions for creating realistic test data
-
-    /// Creates test OTLP logs data
-    fn create_otlp_logs_data() -> OTLPData {
-        let logs_request = ExportLogsServiceRequest::default();
-        OTLPData::Logs(logs_request)
-    }
-
-    /// Creates test OTLP metrics data
-    fn create_otlp_metrics_data() -> OTLPData {
-        let metrics_request = ExportMetricsServiceRequest::default();
-        OTLPData::Metrics(metrics_request)
-    }
-
-    /// Creates test OTLP trace data
-    fn create_otlp_trace_data() -> OTLPData {
-        let trace_request = ExportTraceServiceRequest::default();
-        OTLPData::Traces(trace_request)
-    }
-
-    /// Creates test OTAP arrow logs data
-    fn create_otap_logs_data(batch_id: i32) -> OTAPData {
-        let batch = BatchArrowRecords {
-            batch_id: batch_id.into(),
-            arrow_payloads: vec![],
-            headers: vec![],
-        };
-        OTAPData::ArrowLogs(batch)
-    }
-
-    /// Creates test OTAP arrow metrics data
-    fn create_otap_metrics_data(batch_id: i32) -> OTAPData {
-        let batch = BatchArrowRecords {
-            batch_id: batch_id.into(),
-            arrow_payloads: vec![],
-            headers: vec![],
-        };
-        OTAPData::ArrowMetrics(batch)
-    }
-
-    /// Creates test OTAP arrow traces data
-    fn create_otap_traces_data(batch_id: i32) -> OTAPData {
-        let batch = BatchArrowRecords {
-            batch_id: batch_id.into(),
-            arrow_payloads: vec![],
-            headers: vec![],
-        };
-        OTAPData::ArrowTraces(batch)
-    }
-
-    #[test]
-    fn test_retry_processor_ack_with_otlp_logs() {
-        let runtime = TestRuntime::<OTLPData>::new();
-        let processor = RetryProcessor::new();
-        let config = ProcessorConfig::new("retry_processor_ack_test");
-        let wrapper = ProcessorWrapper::local(processor, &config);
-
-        runtime
-            .set_processor(wrapper)
-            .run_test(|mut ctx| async move {
-                // Send OTLP logs data
-                let otlp_data = create_otlp_logs_data();
-                let message_id = otlp_data.id();
-                ctx.process(Message::PData(otlp_data)).await.unwrap();
-
-                // Send an ACK for the message
-                ctx.process(Message::Control(ControlMsg::Ack { id: message_id }))
-                    .await
-                    .unwrap();
-
-                // Verify the message was forwarded
-                let emitted = ctx.drain_pdata().await;
-                assert_eq!(emitted.len(), 1);
-            })
-            .validate(|_| async {});
-    }
-
-    #[test]
-    fn test_retry_processor_nack_and_retry_with_otap_metrics() {
-        let runtime = TestRuntime::<OTAPData>::new();
-        let config = RetryConfig {
-            max_retries: 2,
-            initial_retry_delay_ms: 10, // Very short for testing
-            max_retry_delay_ms: 100,
-            backoff_multiplier: 2.0,
-            max_pending_messages: 10000,
-            cleanup_interval_secs: 60,
-        };
-        let processor = RetryProcessor::with_config(config);
-        let config = ProcessorConfig::new("retry_processor_nack_test");
-        let wrapper = ProcessorWrapper::local(processor, &config);
-
-        runtime
-            .set_processor(wrapper)
-            .run_test(|mut ctx| async move {
-                // Send OTAP metrics data
-                let otap_data = create_otap_metrics_data(42);
-                let message_id = otap_data.id();
-                ctx.process(Message::PData(otap_data)).await.unwrap();
-
-                // Send a NACK for the message
-                ctx.process(Message::Control(ControlMsg::Nack {
-                    id: message_id,
-                    reason: "Downstream failure".to_string(),
-                }))
-                .await
-                .unwrap();
-
-                // Wait a bit for retry delay
-                tokio::time::sleep(Duration::from_millis(15)).await;
-
-                // Send a timer tick to trigger retry processing
-                ctx.process(Message::Control(ControlMsg::TimerTick {}))
-                    .await
-                    .unwrap();
-
-                // Verify messages were sent (original + retry)
-                let emitted = ctx.drain_pdata().await;
-                assert_eq!(emitted.len(), 2);
-            })
-            .validate(|_| async {});
-    }
-
-    #[test]
-    fn test_retry_processor_max_retries_exceeded_with_otlp_traces() {
-        let runtime = TestRuntime::<OTLPData>::new();
-        let config = RetryConfig {
-            max_retries: 1,
-            initial_retry_delay_ms: 10,
-            max_retry_delay_ms: 100,
-            backoff_multiplier: 2.0,
-            max_pending_messages: 10000,
-            cleanup_interval_secs: 60,
-        };
-        let processor = RetryProcessor::with_config(config);
-        let config = ProcessorConfig::new("retry_processor_max_retries_test");
-        let wrapper = ProcessorWrapper::local(processor, &config);
-
-        runtime
-            .set_processor(wrapper)
-            .run_test(|mut ctx| async move {
-                // Send OTLP trace data
-                let otlp_data = create_otlp_trace_data();
-                let message_id = otlp_data.id();
-                ctx.process(Message::PData(otlp_data)).await.unwrap();
-
-                // Send first NACK
-                ctx.process(Message::Control(ControlMsg::Nack {
-                    id: message_id,
-                    reason: "First failure".to_string(),
-                }))
-                .await
-                .unwrap();
-
-                // Wait and trigger retry
-                tokio::time::sleep(Duration::from_millis(15)).await;
-                ctx.process(Message::Control(ControlMsg::TimerTick {}))
-                    .await
-                    .unwrap();
-
-                // Send second NACK (should exceed max retries)
-                ctx.process(Message::Control(ControlMsg::Nack {
-                    id: message_id,
-                    reason: "Second failure".to_string(),
-                }))
-                .await
-                .unwrap();
-
-                // Trigger retry processing again
-                ctx.process(Message::Control(ControlMsg::TimerTick {}))
-                    .await
-                    .unwrap();
-
-                // Should have original + one retry + shutdown attempt (3 total)
-                // Note: Even though max retries was exceeded, shutdown still attempts to send pending messages
-                let emitted = ctx.drain_pdata().await;
-                assert_eq!(emitted.len(), 3);
-            })
-            .validate(|_| async {});
-    }
-
-    #[test]
-    fn test_retry_processor_config_update_with_otap_traces() {
-        let runtime = TestRuntime::<OTAPData>::new();
-        let processor = RetryProcessor::new();
-        let config = ProcessorConfig::new("retry_processor_config_test");
-        let wrapper = ProcessorWrapper::local(processor, &config);
-
-        runtime
-            .set_processor(wrapper)
-            .run_test(|mut ctx| async move {
-                // Send a config update
-                let new_config = RetryConfig {
-                    max_retries: 5,
-                    initial_retry_delay_ms: 500,
-                    max_retry_delay_ms: 10000,
-                    backoff_multiplier: 1.5,
-                    max_pending_messages: 5000,
-                    cleanup_interval_secs: 30,
-                };
-                let config_json = serde_json::to_value(new_config).unwrap();
-
-                ctx.process(Message::Control(ControlMsg::Config {
-                    config: config_json,
-                }))
-                .await
-                .unwrap();
-
-                // No assertions needed, just verify it doesn't crash
-            })
-            .validate(|_| async {});
-    }
-
-    #[test]
-    fn test_retry_processor_shutdown_with_otap_logs() {
-        let runtime = TestRuntime::<OTAPData>::new();
-        let processor = RetryProcessor::new();
-        let config = ProcessorConfig::new("retry_processor_shutdown_test");
-        let wrapper = ProcessorWrapper::local(processor, &config);
-
-        runtime
-            .set_processor(wrapper)
-            .run_test(|mut ctx| async move {
-                // Send OTAP logs data
-                let otap_data = create_otap_logs_data(123);
-                let message_id = otap_data.id();
-                ctx.process(Message::PData(otap_data)).await.unwrap();
-
-                // Send NACK to put message in retry queue
-                ctx.process(Message::Control(ControlMsg::Nack {
-                    id: message_id,
-                    reason: "Failure".to_string(),
-                }))
-                .await
-                .unwrap();
-
-                // Send shutdown - should attempt to send pending messages
-                ctx.process(Message::Control(ControlMsg::Shutdown {
-                    deadline: Duration::from_secs(5),
-                    reason: "Test shutdown".to_string(),
-                }))
-                .await
-                .unwrap();
-
-                // Should have original message + shutdown attempt
-                let emitted = ctx.drain_pdata().await;
-                assert_eq!(emitted.len(), 2);
-            })
-            .validate(|_| async {});
-    }
-
-    #[test]
-    fn test_retry_processor_different_otlp_data_types() {
-        let runtime = TestRuntime::<OTLPData>::new();
-        let processor = RetryProcessor::new();
-        let config = ProcessorConfig::new("retry_processor_data_types_test");
-        let wrapper = ProcessorWrapper::local(processor, &config);
-
-        runtime
-            .set_processor(wrapper)
-            .run_test(|mut ctx| async move {
-                // Send different OTLP data types
-                let logs_data = create_otlp_logs_data();
-                let metrics_data = create_otlp_metrics_data();
-                let traces_data = create_otlp_trace_data();
-
-                ctx.process(Message::PData(logs_data.clone()))
-                    .await
-                    .unwrap();
-                ctx.process(Message::PData(metrics_data.clone()))
-                    .await
-                    .unwrap();
-                ctx.process(Message::PData(traces_data.clone()))
-                    .await
-                    .unwrap();
-
-                // Verify all messages were forwarded
-                let emitted = ctx.drain_pdata().await;
-                assert_eq!(emitted.len(), 3);
-
-                // Verify each data type has unique IDs
-                let logs_id = logs_data.id();
-                let metrics_id = metrics_data.id();
-                let traces_id = traces_data.id();
-
-                // All IDs should be different
-                assert_ne!(logs_id, metrics_id);
-                assert_ne!(logs_id, traces_id);
-                assert_ne!(metrics_id, traces_id);
-            })
-            .validate(|_| async {});
-    }
-
-    #[test]
-    fn test_retry_processor_different_otap_data_types() {
-        let runtime = TestRuntime::<OTAPData>::new();
-        let processor = RetryProcessor::new();
-        let config = ProcessorConfig::new("retry_processor_otap_data_types_test");
-        let wrapper = ProcessorWrapper::local(processor, &config);
-
-        runtime
-            .set_processor(wrapper)
-            .run_test(|mut ctx| async move {
-                // Send different OTAP data types with unique batch IDs
-                let logs_data = create_otap_logs_data(100);
-                let metrics_data = create_otap_metrics_data(200);
-                let traces_data = create_otap_traces_data(300);
-
-                ctx.process(Message::PData(logs_data.clone()))
-                    .await
-                    .unwrap();
-                ctx.process(Message::PData(metrics_data.clone()))
-                    .await
-                    .unwrap();
-                ctx.process(Message::PData(traces_data.clone()))
-                    .await
-                    .unwrap();
-
-                // Verify all messages were forwarded
-                let emitted = ctx.drain_pdata().await;
-                assert_eq!(emitted.len(), 3);
-
-                // Verify each data type has the expected batch ID as its ID
-                assert_eq!(logs_data.id(), 100);
-                assert_eq!(metrics_data.id(), 200);
-                assert_eq!(traces_data.id(), 300);
-            })
-            .validate(|_| async {});
-    }
-
-    #[test]
-    fn test_backoff_multiplier_calculation() {
-        // Test the exponential backoff calculation directly
-        let config = RetryConfig {
-            max_retries: 5,
-            initial_retry_delay_ms: 1000,
-            max_retry_delay_ms: 30000,
-            backoff_multiplier: 2.0,
-            max_pending_messages: 10000,
-            cleanup_interval_secs: 60,
-        };
-
-        // Manually calculate expected delays for different retry counts
-        // Formula: initial_delay * multiplier^(retry_count - 1)
-
-        // Retry 1: 1000 * 2.0^0 = 1000ms
-        let delay_1 = (config.initial_retry_delay_ms as f64 * config.backoff_multiplier.powi(0))
-            .min(config.max_retry_delay_ms as f64) as u64;
-        assert_eq!(delay_1, 1000);
-
-        // Retry 2: 1000 * 2.0^1 = 2000ms
-        let delay_2 = (config.initial_retry_delay_ms as f64 * config.backoff_multiplier.powi(1))
-            .min(config.max_retry_delay_ms as f64) as u64;
-        assert_eq!(delay_2, 2000);
-
-        // Retry 3: 1000 * 2.0^2 = 4000ms
-        let delay_3 = (config.initial_retry_delay_ms as f64 * config.backoff_multiplier.powi(2))
-            .min(config.max_retry_delay_ms as f64) as u64;
-        assert_eq!(delay_3, 4000);
-
-        // Retry 4: 1000 * 2.0^3 = 8000ms
-        let delay_4 = (config.initial_retry_delay_ms as f64 * config.backoff_multiplier.powi(3))
-            .min(config.max_retry_delay_ms as f64) as u64;
-        assert_eq!(delay_4, 8000);
-
-        // Retry 5: 1000 * 2.0^4 = 16000ms
-        let delay_5 = (config.initial_retry_delay_ms as f64 * config.backoff_multiplier.powi(4))
-            .min(config.max_retry_delay_ms as f64) as u64;
-        assert_eq!(delay_5, 16000);
-
-        // Test that max delay is capped
-        // If we had more retries, they would be capped at max_retry_delay_ms
-        let delay_large = (config.initial_retry_delay_ms as f64
-            * config.backoff_multiplier.powi(10)) // 1000 * 2^10 = 1,024,000
-        .min(config.max_retry_delay_ms as f64) as u64;
-        assert_eq!(delay_large, 30000); // Should be capped at max_retry_delay_ms
-    }
-
-    #[test]
-    fn test_otlp_and_otap_id_generation() {
-        // Test ID generation for real telemetry data types
-        let otlp_logs = create_otlp_logs_data();
-        let otlp_metrics = create_otlp_metrics_data();
-        let otlp_traces = create_otlp_trace_data();
-
-        let otap_logs = create_otap_logs_data(1001);
-        let otap_metrics = create_otap_metrics_data(1002);
-        let otap_traces = create_otap_traces_data(1003);
-
-        // Test that IDs are deterministic
-        let otlp_logs_2 = create_otlp_logs_data();
-        assert_eq!(otlp_logs.id(), otlp_logs_2.id());
-
-        // Test that different OTLP types have different IDs
-        assert_ne!(otlp_logs.id(), otlp_metrics.id());
-        assert_ne!(otlp_logs.id(), otlp_traces.id());
-        assert_ne!(otlp_metrics.id(), otlp_traces.id());
-
-        // Test that OTAP IDs match batch IDs
-        assert_eq!(otap_logs.id(), 1001);
-        assert_eq!(otap_metrics.id(), 1002);
-        assert_eq!(otap_traces.id(), 1003);
-
-        // Test deadlines (should be None for both types currently)
-        assert!(otlp_logs.deadline().is_none());
-        assert!(otap_logs.deadline().is_none());
-
-        println!("Real telemetry data types work correctly for retry processor testing!");
-    }
-
-    #[test]
-    fn test_backoff_multiplier_different_values() {
-        // Test with different backoff multipliers
-
-        // Test with multiplier 1.5
-        let config_1_5 = RetryConfig {
-            max_retries: 3,
-            initial_retry_delay_ms: 1000,
-            max_retry_delay_ms: 10000,
-            backoff_multiplier: 1.5,
-            max_pending_messages: 10000,
-            cleanup_interval_secs: 60,
-        };
-
-        let delay_1 = (config_1_5.initial_retry_delay_ms as f64
-            * config_1_5.backoff_multiplier.powi(0)) as u64; // 1000ms
-        let delay_2 = (config_1_5.initial_retry_delay_ms as f64
-            * config_1_5.backoff_multiplier.powi(1)) as u64; // 1500ms
-        let delay_3 = (config_1_5.initial_retry_delay_ms as f64
-            * config_1_5.backoff_multiplier.powi(2)) as u64; // 2250ms
-
-        assert_eq!(delay_1, 1000);
-        assert_eq!(delay_2, 1500);
-        assert_eq!(delay_3, 2250);
-
-        // Test with multiplier 3.0 (more aggressive)
-        let config_3_0 = RetryConfig {
-            max_retries: 3,
-            initial_retry_delay_ms: 500,
-            max_retry_delay_ms: 20000,
-            backoff_multiplier: 3.0,
-            max_pending_messages: 10000,
-            cleanup_interval_secs: 60,
-        };
-
-        let delay_1 = (config_3_0.initial_retry_delay_ms as f64
-            * config_3_0.backoff_multiplier.powi(0)) as u64; // 500ms
-        let delay_2 = (config_3_0.initial_retry_delay_ms as f64
-            * config_3_0.backoff_multiplier.powi(1)) as u64; // 1500ms
-        let delay_3 = (config_3_0.initial_retry_delay_ms as f64
-            * config_3_0.backoff_multiplier.powi(2)) as u64; // 4500ms
-
-        assert_eq!(delay_1, 500);
-        assert_eq!(delay_2, 1500);
-        assert_eq!(delay_3, 4500);
-    }
-
-    #[test]
-    fn test_retry_queue_memory_limit_with_otlp_data() {
-        let runtime = TestRuntime::<OTLPData>::new();
-        let config = RetryConfig {
-            max_retries: 3,
-            initial_retry_delay_ms: 1000,
-            max_retry_delay_ms: 30000,
-            backoff_multiplier: 2.0,
-            max_pending_messages: 2, // Very small limit for testing
-            cleanup_interval_secs: 60,
-        };
-        let processor = RetryProcessor::with_config(config);
-        let config = ProcessorConfig::new("memory_limit_test");
-        let wrapper = ProcessorWrapper::local(processor, &config);
-
-        runtime
-            .set_processor(wrapper)
-            .run_test(|mut ctx| async move {
-                // Send messages up to the limit
-                let otlp_data_1 = create_otlp_logs_data();
-                let otlp_data_2 = create_otlp_metrics_data();
-                let otlp_data_3 = create_otlp_trace_data(); // This should trigger the limit
-
-                ctx.process(Message::PData(otlp_data_1)).await.unwrap();
-                ctx.process(Message::PData(otlp_data_2)).await.unwrap();
-                ctx.process(Message::PData(otlp_data_3)).await.unwrap(); // Should still work but log warning
-
-                // All messages should still be forwarded even if queue is full
-                let emitted = ctx.drain_pdata().await;
-                assert_eq!(emitted.len(), 3);
-            })
-            .validate(|_| async {});
-    }
-
-    #[test]
-    fn test_expired_message_cleanup() {
-        // This test verifies that the cleanup functionality exists and can be configured
-        // A more comprehensive test would require access to the processor's internal state
-        let config = RetryConfig {
-            max_retries: 3,
-            initial_retry_delay_ms: 1000,
-            max_retry_delay_ms: 30000,
-            backoff_multiplier: 2.0,
-            max_pending_messages: 10000,
-            cleanup_interval_secs: 1, // Very short cleanup interval for testing
-        };
-        let processor = RetryProcessor::<OTLPData>::with_config(config);
-
-        // Verify the processor was created with the correct cleanup interval
-        assert_eq!(processor.config.cleanup_interval_secs, 1);
-
-        // Note: Testing the actual cleanup behavior would require either:
-        // 1. Exposing internal state (not ideal for encapsulation)
-        // 2. Integration tests with actual message expiration
-        // 3. Dependency injection for time (more complex)
-        // For now, we test that the configuration is properly set
-    }
-
-    #[test]
-    fn test_memory_safety_config_validation() {
-        // Test that configuration includes memory safety parameters
-        let default_config = RetryConfig::default();
-        assert_eq!(default_config.max_pending_messages, 10000);
-        assert_eq!(default_config.cleanup_interval_secs, 60);
-
-        // Test custom config
-        let custom_config = RetryConfig {
-            max_retries: 5,
-            initial_retry_delay_ms: 2000,
-            max_retry_delay_ms: 60000,
-            backoff_multiplier: 1.5,
-            max_pending_messages: 5000,
-            cleanup_interval_secs: 30,
-        };
-
-        assert_eq!(custom_config.max_pending_messages, 5000);
-        assert_eq!(custom_config.cleanup_interval_secs, 30);
-
-        // Verify the processor accepts the config
-        let processor = RetryProcessor::<OTLPData>::with_config(custom_config);
-        assert_eq!(processor.config.max_pending_messages, 5000);
-    }
-}
